Crate preflight

Utilities to create and test hardware agnostic flight systems with little friction

This preflight crate goes hand and hand with the cargo_preflight cargo subcommand. Alone, this crate will produce a stable, plugable interface for a flight computer to link to. When paired with the cargo_preflight command, users are able to run the flight systems through rigorous simulations and tests to verify their integrity.

Re-exports

pub use uom;

Modules

abi
This module should never be consumed directly. The ABI used to communicate between the preflight runner and the avionics.

Structs

Guidance
A hardware agnostic guidance signal

Sensors
Generic sensor values that are collected from the flight hardware which would be useful to calculate position and velocity.

ThrustVector
A call for thrust vectoring hardware to produce a thrust at the given vector

Vector3
A vector representing a quantity in 3 dimensional space

Enums

AbortCause
Underlying cause for an abort

Control
Various control signals that the avionics can produce

Traits

Avionics
Hardware agnostic avionics system

Type Definitions

Quantity
Generic uom quantity using f32 as the storage type

Attribute Macros

avionics_harness
Harness to connect hardware agnostic flight systems to firmware or to the preflight_cargo utility