Module preflight::abi

This module should never be consumed directly. The ABI used to communicate between the preflight runner and the avionics.

This module's contents are only made public for use by code generated by the avionics_harness macro

Type Definitions

AvionicsGuide
The entry point into the user's avionics

PanicCallback
Callback that is called when the driven avionics panics.

SetPanicCallback
Function exposed to the preflight runner that can set the PanicCallback